+++ pyfor/collection.py
@@ -1,17 +1,16 @@
 import os
 import laspy
-import pandas as pd
+import numpy as np
 import pyfor
 import geopandas as gpd
+import pandas as pd
+import laxpy
 
-class Indexer:
-    """
-    Internal class used to index a directory of las files.
-    """
 
 class CloudDataFrame(gpd.GeoDataFrame):
     """
-    Implements a data frame structure for processing and managing multiple cloud objects.
+    Implements a data frame structure for processing and managing multiple :class:`.Cloud` objects. It is recommended \
+    to initialize using the :func:`.from_dir` function.
     """
     def __init__(self, *args, **kwargs):
         super(CloudDataFrame, self).__init__(*args, **kwargs)
@@ -21,10 +20,10 @@ class CloudDataFrame(gpd.GeoDataFrame):
             self.set_geometry("bounding_box", inplace=True)
 
     @classmethod
-    def from_dir(cls, las_dir, n_jobs=1, get_bounding_boxes = True):
+    def _from_dir(cls, las_dir, n_jobs=1, get_bounding_boxes=True):
         """
         Wrapped function for producing a CloudDataFrame from a directory of las files.
-        :param las_dir:
+        :param las_dir: A directory of .las or .laz files.
         :param n_jobs: The number of threads used to construct information about the CloudDataFrame.
         :param get_bounding_boxes: If True, builds the bounding boxes for each las tile by manually reading in
         the file and computing the bounding box. For very large collections this may be computationally costly, and
@@ -32,6 +31,7 @@ class CloudDataFrame(gpd.GeoDataFrame):
         :return:
         """
         las_path_init = [[os.path.join(root, file) for file in files] for root, dirs, files in os.walk(las_dir)][0]
+        las_path_init = [las_path for las_path in las_path_init if las_path.endswith('.las') or las_path.endswith('.laz')]
         cdf = CloudDataFrame({'las_path': las_path_init})
         cdf.n_threads = n_jobs
 
@@ -43,14 +43,13 @@ class CloudDataFrame(gpd.GeoDataFrame):
     def set_index(self, *args):
         return CloudDataFrame(super(CloudDataFrame, self).set_index(*args))
 
-    def par_apply(self, func, column='las_path', buffer_distance = 0, *args):
+    def par_apply(self, func, column='las_path', buffer_distance=0, *args):
         """
-        Apply a function to each las path. Allows for parallelization using the n_jobs argument. This is achieved \
-        via joblib Parallel and delayed.
+        Apply a function to each item in `column`. Allows for parallelization using the n_jobs argument. This is \
+        achieved via :class:`joblib.Parallel` and :func:`joblib.delayed`.
 
-        :param func: The user defined function, must accept a single argument, the path of the las file.
-        :param n_jobs: The number of threads to spawn, default of 1.
-        :param column: The column to apply on, will be the first argument to func
+        :param func: The user defined function, must accept as an argument the value of each row of `column`.
+        :param column: The column to apply on, will be the first argument to func.
         :param buffer_distance: The distance to buffer and aggregate each tile.
         :param *args: Further arguments to `func`
         """
@@ -59,9 +58,7 @@ class CloudDataFrame(gpd.GeoDataFrame):
             self._buffer(buffer_distance)
             for i, geom in enumerate(self["bounding_box"]):
                 intersecting = self._get_intersecting(i)
-
                 clip_geom = self['buffered_bounding_box'].iloc[i]
-                print(clip_geom)
                 parent_cloud = pyfor.cloud.Cloud(self["las_path"].iloc[i])
                 for path in intersecting["las_path"]:
                     adjacent_cloud = pyfor.cloud.Cloud(path)
@@ -70,8 +67,6 @@ class CloudDataFrame(gpd.GeoDataFrame):
         output = Parallel(n_jobs=self.n_threads)(delayed(func)(plot_path, *args) for plot_path in self[column])
         return output
 
-    # TODO Many of these _functions are redundant due to a bug in joblib that prevents lambda functions
-    # once this bug is fixed these functions can be drastically simplified and aggregated.
     def _get_bounding_box(self, las_path):
         """
         Vectorized function to get a bounding box from an individual las path.
@@ -84,17 +79,10 @@ class CloudDataFrame(gpd.GeoDataFrame):
         min_y, max_y = pc.header.min[1], pc.header.max[1]
         return((min_x, max_x, min_y, max_y))
 
-    def _get_bounding_boxes(self):
-        """
-        Retrieves a bounding box for each path in las path.
-        :return:
-        """
-        return self.par_apply(self._get_bounding_box, column="las_path")
-
     def _build_polygons(self):
         """Builds the shapely polygons of the bounding boxes and adds them to self.data"""
         from shapely.geometry import Polygon
-        bboxes = self._get_bounding_boxes()
+        bboxes = self.par_apply(self._get_bounding_box, column='las_path')
         self["bounding_box"] = [Polygon(((bbox[0], bbox[2]), (bbox[1], bbox[2]),
                                            (bbox[1], bbox[3]), (bbox[0], bbox[3]))) for bbox in bboxes]
         self.set_geometry("bounding_box", inplace = True)
@@ -113,7 +101,6 @@ class CloudDataFrame(gpd.GeoDataFrame):
         intersect_cdf.n_threads = self.n_threads
         return intersect_cdf
 
-
     def _buffer(self, distance, in_place = True):
         """
         Buffers the CloudDataFrame geometries.
@@ -130,21 +117,170 @@ class CloudDataFrame(gpd.GeoDataFrame):
         cdf.set_geometry("bounding_box", inplace=True)
         return cdf
 
-    def clip(self):
+    def plot(self, **kwargs):
         """
-        Clips the CloudDataFrame with the supplied geometries.
-        :return:
+        Plots the bounding boxes of the Cloud objects.
+
+        :param **kwargs: Keyword arguments to :meth:`geopandas.GeoDataFrame.plot`.
         """
-        pass
-
-
-    def plot(self, *args):
-        """Plots the bounding boxes of the Cloud objects"""
-        plot = super(CloudDataFrame, self).plot(*args)
+        plot = super(CloudDataFrame, self).plot(**kwargs)
         plot.figure.show()
 
+    @property
+    def bounding_box(self):
+        """Retrieves the bounding box for the entire collection. As a tuple (minx, muny, maxx, maxy)"""
+        minx, miny, maxx, maxy = [i.bounds[0] for i in self['bounding_box']], [i.bounds[1] for i in self['bounding_box']], \
+                                 [i.bounds[2] for i in self['bounding_box']], [i.bounds[3] for i in self['bounding_box']]
+        col_bbox = np.min(minx), np.min(miny), np.max(maxx), np.max(maxy)
+        return col_bbox
 
-def from_dir(las_dir, n_jobs=1):
+    def retile(self, out_dir, verbose = False):
+        """
+        Retiles the collection into `out_dir`. This is a simplified retiling function that splits each tile into \
+        quadrants and writes these new quadrants to disk. The written files will be the original file name with an index
+        appended. 0 is the bottom left quadrant, 1 is the bottom right, 2 is the top left and 3 is the top right.
+        """
+        from shapely.geometry import Polygon
+
+        for index, row in self.iterrows():
+            # Build the quadrant geometries, this is defined by the following six values
+            larger_cell = row['bounding_box'].bounds
+
+            x0, y0 = larger_cell[0], larger_cell[1]
+            x2, y2 = larger_cell[2], larger_cell[3]
+            x1 = ((larger_cell[2] - larger_cell[0]) / 2) + x0
+            y1 = ((larger_cell[3] - larger_cell[1]) / 2) + y0
+
+            # Create the geometries
+            bottom_left = Polygon([(x0, y0), (x0, y1), (x1, y1), (x1, y0)])
+            bottom_right = Polygon([(x1, y0), (x1, y1), (x2, y1), (x2, y0)])
+            top_left = Polygon([(x0, y1), (x0, y2), (x1, y2), (x1, y1)])
+            top_right = Polygon([(x1, y1), (x1, y2), (x2, y2), (x2, y1)])
+
+            quadrants = [bottom_left, bottom_right, top_left, top_right]
+
+
+            larger_cloud = pyfor.cloud.Cloud(row['las_path'])
+
+            for i, quad in enumerate(quadrants):
+                clipped = larger_cloud.clip(quad)
+                clipped.write(os.path.join(out_dir, '{}_{}.las'.format(larger_cloud.name, i)))
+
+    def _retile2(self, width, height, dir):
+        """
+        Retiles the collection and writes the new tiles to the directory defined in `dir`.
+
+        :param width: The width of the new tiles.
+        :param height: The height of the new tiles
+        :param dir: The directory to write the new tiles.
+        """
+        # TODO Handle "edge" smaller tiles that straddle more than one larger tile
+        from shapely.geometry import MultiLineString
+        from shapely.ops import polygonize
+        import warnings
+
+        warnings.warn('_retile2 is in development, use at your own risk.', UserWarning)
+
+        colbbox = self.bounding_box
+        x = np.arange(colbbox[0], colbbox[2], width)
+        y = np.arange(colbbox[1], colbbox[3], height)
+
+        hlines = [((x1, yi), (x2, yi)) for x1, x2 in zip(x[:-1], x[1:]) for yi in y]
+        vlines = [((xi, y1), (xi, y2)) for y1, y2 in zip(y[:-1], y[1:]) for xi in x]
+
+        grids = gpd.GeoSeries(polygonize(MultiLineString(hlines + vlines)))
+
+        # Iterate through each larger tile and find the intersecting smaller tiles
+        # TODO better way for this?
+        for index, row in self.iterrows():
+            # Find the set of smaller tiles that intersect with the larger
+            # Load the larger cell into memory
+            larger_cell = pyfor.cloud.Cloud(row['las_path'])
+            for i, smaller_cell in enumerate(grids[grids.intersects(row['bounding_box'])]):
+                pc = larger_cell.clip(smaller_cell)
+                pc.write(os.path.join(dir, '{}_{}{}'.format(larger_cell.name, i, larger_cell.extension)))
+
+    def create_index(self):
+        """
+        For each file in the collection, creates `.lax` files for spatial indexing using the default values.
+        """
+        for las_path in self['las_path']:
+            laxpy.file.init_lax(las_path)
+
+    def _index_las(self, las_path):
+        """
+        Checks if an equivalent `.lax` file exists. If so, creates a laxpy.IndexedLAS object, otherwise an error is thrown.
+        """
+        lax_path = las_path[:-1] + 'x'
+
+        if os.path.isfile(lax_path):
+            return laxpy.IndexedLAS(las_path)
+        else:
+            raise FileNotFoundError('There is no equivalent .lax file for this .las file.')
+
+    def clip(self, polygons, path, poly_names=None):
+        """
+        A collection-level clipping method. This function is meant for efficient querying across the study area using \
+        a set of polygons. This method requires the presence of `.lax` files in the collection directory. To generate \
+        these `.lax` files please use :meth:`.create_index` first. Each polygon will be clipped and written to the \
+        specified `path`.
+
+        :param polygons: Either a list or :class:`geopandas.GeoSeries` of shapely polygons.
+        :param path: The output path of the clip.
+        :param poly_names: A list of polygon names to use when writing to file.
+        """
+        # TODO currently does not take advantage of multi-threading
+        # TODO also a bit long, may be best to break up
+        if ~hasattr(polygons, '__iter__') and type(polygons) != gpd.GeoSeries:
+            polygons = [polygons]
+
+        head, tail = os.path.split(path)
+
+        # Which tiles do I need to make an index for?
+        # It could  be the case that the input polys intersect with the same tile, but are checked out of order
+        # Building this dict requires a bit of overhead, but is more memory efficient in the worst case
+        intersected_tiles = {}
+        for ix, row in self.iterrows():
+            tile_bbox, las_path = row['bounding_box'], row['las_path']
+            for poly in polygons:
+                if tile_bbox.intersects(poly):
+                    if las_path in intersected_tiles:
+                        intersected_tiles[las_path].append(poly)
+                    else:
+                        intersected_tiles[las_path] = [poly]
+
+        # Which polygons have which parents?
+        parents = {}
+        for i, poly in enumerate(polygons):
+            poly_parents = []
+            for las_path, poly_list in intersected_tiles.items():
+                if poly in poly_list:
+                    poly_parents.append(las_path)
+                parents[i] = poly_parents
+
+        # For each polygon (i.e. each index in parents) construct the clipped point cloud.
+        # maybe this could just be done in the chunk above?
+        for poly_index, parent_list in parents.items():
+            poly = polygons[poly_index]
+            indexed_parents = [self._index_las(parent_path) for parent_path in parent_list]
+            header = indexed_parents[0].header
+            # TODO This is slow, but should be addressed upstream in laxpy, especially _scale_points
+            parent_points = pd.concat([pd.DataFrame.from_records(parent.query_polygon(poly, scale=True)) for parent in indexed_parents])
+
+            print('Clipping polygon {} of {}'.format(poly_index + 1, len(polygons)))
+            pc = pyfor.cloud.Cloud(pyfor.cloud.LASData(parent_points, header))
+
+            if poly_names is not None:
+                out_path = head + os.path.sep + str(poly_names[poly_index]) + '.las'
+            else:
+                out_path = head + os.path.sep + str(poly_index) + '.las'
+
+            print('Writing to {}'.format(out_path))
+            pc.write(out_path)
+
+
+
+def from_dir(las_dir, **kwargs):
     """
     Constructs a CloudDataFrame from a directory of las files.
 
@@ -152,12 +288,5 @@ def from_dir(las_dir, n_jobs=1):
     :return: A CloudDataFrame constructed from the directory of las files.
     """
 
-    return CloudDataFrame.from_dir(las_dir, n_jobs= n_jobs)
+    return CloudDataFrame._from_dir(las_dir, **kwargs)
 
-
-def stitch_clouds(collection):
-    """
-    Holder function for some ideas.
-    :return:
-
-    """